requirements.txt
# Project dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
openai==1.3.5
beautifulsoup4==4.12.2
requests==2.31.0
pytest==7.4.3
google-adk
readability-lxml==0.8.1
langdetect==1.0.9
youtube-transcript-api==0.6.2
